chore: fix imports move is-billing-plan-changes to myoffer page utils
