+++ src/components/pages/storage/myoffers/StorageEditOfferPage.tsx
@@ -18,11 +18,12 @@ import { useHistory } from 'react-router-dom'
 import ROUTES from 'routes'
 import Logger from 'utils/Logger'
 import TransactionInProgressPanel from 'components/organisms/TransactionInProgressPanel'
-import { isBillingPlansChanged, transformOfferDataForContract } from 'contracts/storage/utils'
+import { transformOfferDataForContract } from 'contracts/storage/utils'
 import { SupportedTokens } from 'contracts/interfaces'
 import { StorageOffer } from 'models/marketItems/StorageItem'
 import Web3 from 'web3'
-import { UNIT_PREFIX_POW2 } from '../../../../utils/utils'
+import { UNIT_PREFIX_POW2 } from 'utils/utils'
+import { isBillingPlansChanged } from 'components/pages/storage/myoffers/utils'
 
 const logger = Logger.getInstance()
