+++ src/contracts/storage/utils.ts
@@ -108,19 +108,3 @@ export const transformOfferDataForContract = (
     tokens,
   }
 }
-
-export const isBillingPlansChanged = (
-  current: StorageBillingPlan[],
-  previous: StorageBillingPlan[],
-): boolean => {
-  if (current.length !== previous.length) {
-    return true
-  }
-  return previous.some(
-    (plan) => !current.find(
-      (p) => p.currency === plan.currency
-              && p.period === plan.period
-              && p.price.toString() === plan.price.toString(),
-    ),
-  )
-}
